backend/pyproject.toml
[project]
name = "taskwise-backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "sqlalchemy>=2.0.42",
    "uvicorn>=0.35.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "crudadmin>=0.4.2",
    "itsdangerous>=2.2.0",
    "webauthn>=2.7.0",
]
